+++ pkg/codegen/go/gen_program.go
@@ -24,6 +24,7 @@ type generator struct {
 	jsonTempSpiller     *jsonSpiller
 	ternaryTempSpiller  *tempSpiller
 	readDirTempSpiller  *readDirSpiller
+	splatSpiller        *splatSpiller
 	scopeTraversalRoots codegen.StringSet
 }
 
@@ -36,6 +37,7 @@ func GenerateProgram(program *hcl2.Program) (map[string][]byte, hcl.Diagnostics,
 		jsonTempSpiller:     &jsonSpiller{},
 		ternaryTempSpiller:  &tempSpiller{},
 		readDirTempSpiller:  &readDirSpiller{},
+		splatSpiller:        &splatSpiller{},
 		scopeTraversalRoots: codegen.NewStringSet(),
 	}
 
@@ -183,7 +185,7 @@ func (g *generator) genResource(w io.Writer, r *hcl2.Resource) {
 	}
 
 	instantiate := func(varName, resourceName string) {
-		if g.scopeTraversalRoots.Has(varName) {
+		if g.scopeTraversalRoots.Has(varName) || strings.HasPrefix(varName, "_") {
 			g.Fgenf(w, "%s, err := %s.New%s(ctx, %s, ", varName, mod, typ, resourceName)
 		} else {
 			g.Fgenf(w, "_, err = %s.New%s(ctx, %s, ", mod, typ, resourceName)
@@ -210,8 +212,11 @@ func (g *generator) genResource(w io.Writer, r *hcl2.Resource) {
 		rangeExpr, temps := g.lowerExpression(r.Options.Range, rangeType, false)
 		g.genTemps(w, temps)
 
-		g.Fgenf(w, "for i0, val0 := range %.v {\n", rangeExpr)
-		instantiate("_", fmt.Sprintf("\"%s-\"+ string(i0)", resName))
+		g.Fgenf(w, "var %s []*%s.%s\n", resName, mod, typ)
+
+		g.Fgenf(w, "for key0, val0 := range %.v {\n", rangeExpr)
+		instantiate("_res", fmt.Sprintf(`fmt.Sprintf("%s-%%v", key0)`, resName))
+		g.Fgenf(w, "%s = append(%s, _res)\n", resName, resName)
 		g.Fgenf(w, "}\n")
 
 	} else {
@@ -286,10 +291,16 @@ func (g *generator) genTempsMultiReturn(w io.Writer, temps []interface{}, zeroVa
 			g.Fgenf(w, "}\n")
 			namesVar := fmt.Sprintf("fileNames%s", tmpSuffix)
 			g.Fgenf(w, "%s := make([]string, len(%s))\n", namesVar, t.Name)
-			iVar := fmt.Sprintf("i%s", tmpSuffix)
+			iVar := fmt.Sprintf("key%s", tmpSuffix)
 			valVar := fmt.Sprintf("val%s", tmpSuffix)
 			g.Fgenf(w, "for %s, %s := range %s {\n", iVar, valVar, t.Name)
 			g.Fgenf(w, "%s[%s] = %s.Name()\n", namesVar, iVar, valVar)
+			g.Fgenf(w, "}\n")
+		case *splatTemp:
+			argTyp := argumentTypeName(t.Value.Each, t.Value.Each.Type(), false)
+			g.Fgenf(w, "var %s []%s\n", t.Name, argTyp)
+			g.Fgenf(w, "for _, val0 := range %.v {\n", t.Value.Source)
+			g.Fgenf(w, "%s = append(%s, %.v)\n", t.Name, t.Name, t.Value.Each)
 			g.Fgenf(w, "}\n")
 		default:
 			contract.Failf("unexpected temp type: %v", t)